requirements.txt
python-gnupg
unidecode
